+++ spec/rspec/matchers/built_in/eql_spec.rb
@@ -1,7 +1,7 @@
 module RSpec
   module Matchers
     RSpec.describe "eql" do
-      it_behaves_like "an RSpec matcher", :valid_value => 1, :invalid_value => 2 do
+      it_behaves_like "an RSpec value matcher", :valid_value => 1, :invalid_value => 2 do
         let(:matcher) { eql(1) }
       end