+++ spec/rspec/matchers/built_in/has_spec.rb
@@ -1,6 +1,6 @@
 RSpec.describe "expect(...).to have_sym(*args)" do
-  it_behaves_like "an RSpec matcher", :valid_value => { :a => 1 },
-                                      :invalid_value => {} do
+  it_behaves_like "an RSpec value matcher", :valid_value => { :a => 1 },
+                                            :invalid_value => {} do
     let(:matcher) { have_key(:a) }
   end