+++ spec/rspec/matchers/built_in/respond_to_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe "expect(...).to respond_to(:sym)" do
-  it_behaves_like "an RSpec matcher", :valid_value => "s", :invalid_value => 5 do
+  it_behaves_like "an RSpec value matcher", :valid_value => "s", :invalid_value => 5 do
     let(:matcher) { respond_to(:upcase) }
   end