+++ spec/rspec/matchers/built_in/start_and_end_with_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe "expect(...).to start_with" do
-  it_behaves_like "an RSpec matcher", :valid_value => "ab", :invalid_value => "bc" do
+  it_behaves_like "an RSpec value matcher", :valid_value => "ab", :invalid_value => "bc" do
     let(:matcher) { start_with("a") }
   end
 
@@ -207,7 +207,7 @@ RSpec.describe "expect(...).not_to start_with" do
 end
 
 RSpec.describe "expect(...).to end_with" do
-  it_behaves_like "an RSpec matcher", :valid_value => "ab", :invalid_value => "bc" do
+  it_behaves_like "an RSpec value matcher", :valid_value => "ab", :invalid_value => "bc" do
     let(:matcher) { end_with("b") }
   end