+++ spec/spec_helper.rb
@@ -5,7 +5,7 @@ RSpec::Support::Spec.setup_simplecov do
   minimum_coverage 97
 end
 
-Dir['./spec/support/**/*'].each do |f|
+Dir['./spec/support/**/*.rb'].each do |f|
   require f.sub(%r|\./spec/|, '')
 end